requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-scd4x==1.4.5
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.76.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
gpiod==2.2.3
gpiodevice==0.0.5
idna==3.10
inky==2.0.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
pyftdi==0.56.0
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyusb==1.3.1
requests==2.32.3
six==1.17.0
smbus2==0.5.0
spidev==3.6
sysv_ipc==1.1.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.3.0
